Remove all the styles
