+++ package.json
@@ -19,7 +19,7 @@
   "bundlesize": [
     {
       "path": "dist/nanocal.min.js",
-      "maxSize": "4.5 kB"
+      "maxSize": "3.7 kB"
     }
   ],
   "dependencies": {
